mimxrt: Change _boot.py to cater for MSC support.
As a benefit, it will not destroy an existing FAT partition.

+++ ports/mimxrt/modules/_boot.py
@@ -8,13 +8,45 @@ import sys
 import mimxrt
 from machine import Pin
 
+FS_UNDEF = const(0)
+FS_LITTLEFS = const(1)
+FS_FAT = const(2)
+
+def fs_type(bdev):
+    b = bytearray(512)
+    bdev.readblocks(0, b)
+    if b[40:48] == b"littlefs" or b[8:16] == b"littlefs":
+        return FS_LITTLEFS
+    elif b[510:512] == b'\x55\xaa':
+        return FS_FAT
+    else:
+        return FS_UNDEF
+
 bdev = mimxrt.Flash()
-try:
-    fs = vfs.VfsLfs2(bdev, progsize=256)
-except:
-    vfs.VfsLfs2.mkfs(bdev, progsize=256)
-    fs = vfs.VfsLfs2(bdev, progsize=256)
-vfs.mount(fs, "/flash")
+
+# In case of MSC support mount as FAT
+# Create a FAT FS is needed
+if hasattr(mimxrt, "MSC"):
+    try:
+        vfs = vfs.VfsFat(bdev)
+        os.mount(vfs, "/flash")
+    except:
+        vfs.VfsFat.mkfs(bdev)
+        vfs = vfs.VfsFat(bdev)
+        os.mount(vfs, "/flash")
+# otherwise analyze the boot sector an mount accordingly
+# if no valid boot sector, creat a LFS file system
+else:
+    fs = fs_type(bdev)
+    if fs == FS_LITTLEFS:
+        vfs = vfs.VfsLfs2(bdev, progsize=256)
+    elif fs == FS_FAT:
+        vfs = vfs.VfsFat(bdev)
+    else:
+        vfs.VfsLfs2.mkfs(bdev, progsize=256)
+        vfs = vfs.VfsLfs2(bdev, progsize=256)
+    os.mount(vfs, "/flash")
+
 os.chdir("/flash")
 sys.path.append("/flash")
 sys.path.append("/flash/lib")
